ar71xx: add device support for COMFAST WR615N
COMFAST CF-WR615N likes a the clone version of TL-WR841N-V9, but
it use a different SoC.
There is an external watchdoy connected to GPIO13, set it to high
would disable it.
There is the three colors LED, the pin map is following here.
GPIO0 -> Blue
GPIO2 -> RED
GPIO3 -> Green
Set those pins to high would light the specific color.